:art: (376): put table logic in MeetingDataTable
