+++ mcr-frontend/src/components.d.ts
@@ -19,15 +19,13 @@ declare module 'vue' {
     BotDisconnecting: typeof import('./components/meeting/transcription/states/BotDisconnecting.vue')['default']
     CaptureVisioInProgress: typeof import('./components/meeting/transcription/states/CaptureVisioInProgress.vue')['default']
     ComuMeetingForm: typeof import('./components/meeting/visio-modal/ComuMeetingForm.vue')['default']
-    copy: typeof import('./components/meeting/table/DataTableRows.vue')['default']
+    copy: typeof import('./components/meeting/table/DataTableCellsAction copy.vue')['default']
     CreateMeetingModal: typeof import('./components/meeting/modals/CreateMeetingModal.vue')['default']
     CreateVisioMeetingForm: typeof import('./components/meeting/CreateVisioMeetingForm.vue')['default']
     CreateVisioMeetingModal: typeof import('./components/meeting/modals/CreateVisioMeetingModal.vue')['default']
     DataCellAction: typeof import('./components/meeting/table/cells/DataCellAction.vue')['default']
     DataCellMeeting: typeof import('./components/meeting/table/cells/DataCellMeeting.vue')['default']
     DataTableCellsAction: typeof import('./components/meeting/table/DataTableCellsAction.vue')['default']
-    DataTableRows: typeof import('./components/meeting/table/DataTableRows.vue')['default']
-    DataTableRowsAction: typeof import('./components/meeting/table/DataTableCellsAction.vue')['default']
     DeleteMeetingModal: typeof import('./components/meeting/modals/DeleteMeetingModal.vue')['default']
     DsfrAlert: typeof import('@gouvminint/vue-dsfr')['DsfrAlert']
     DsfrButton: typeof import('@gouvminint/vue-dsfr')['DsfrButton']
@@ -61,6 +59,7 @@ declare module 'vue' {
     ImportMeetingProgressBar: typeof import('./components/meeting/ImportMeetingProgressBar.vue')['default']
     ImportPending: typeof import('./components/meeting/transcription/states/ImportPending.vue')['default']
     MeetingFrontMatter: typeof import('./components/meeting/MeetingFrontMatter.vue')['default']
+    MeetingsDataTable: typeof import('./components/meeting/table/MeetingsDataTable.vue')['default']
     PageFrontMatter: typeof import('./components/core/PageFrontMatter.vue')['default']
     PageFrontMatterV2: typeof import('./components/core/PageFrontMatterV2.vue')['default']
     ProgressBar: typeof import('./components/core/ProgressBar.vue')['default']
