+++ mcr-frontend/src/locales/fr.json
@@ -121,8 +121,8 @@
         "title": "Titre de réunion",
         "transcription": "Transcription",
         "report": {
-          "label": "Compte-rendu", 
-          "tooltip": "La génération du compte rendu est disponible une fois la transcription prête. Vous pouvez la lancer depuis la fiche réunion, ou la configurer en amont." 
+          "label": "Compte-rendu",
+          "tooltip": "La génération du compte rendu est disponible une fois la transcription prête. Vous pouvez la lancer depuis la fiche réunion, ou la configurer en amont."
         },
         "actions": "Actions"
       }
